move model validators next to models
