+++ src/Halcyon.Api/Models/Account/ResetPasswordRequest.cs
@@ -1,4 +1,6 @@
-﻿namespace Halcyon.Api.Models.Account
+﻿using FluentValidation;
+
+namespace Halcyon.Api.Models.Account
 {
     public class ResetPasswordRequest
     {
@@ -7,5 +9,15 @@
         public string EmailAddress { get; set; }
 
         public string NewPassword { get; set; }
+    }
+
+    public class ResetPasswordRequestValidator : AbstractValidator<ResetPasswordRequest>
+    {
+        public ResetPasswordRequestValidator()
+        {
+            RuleFor(x => x.Token).NotEmpty();
+            RuleFor(x => x.EmailAddress).NotEmpty().EmailAddress().WithName("Email Address");
+            RuleFor(x => x.NewPassword).NotEmpty().MinimumLength(8).MaximumLength(50).WithName("New Password");
+        }
     }
 }
