+++ src/Halcyon.Api/Models/Manage/ChangePasswordRequest.cs
@@ -1,3 +1,5 @@
+using FluentValidation;
+
 namespace Halcyon.Api.Models.Manage
 {
     public class ChangePasswordRequest : UpdateRequest
@@ -5,5 +7,14 @@ namespace Halcyon.Api.Models.Manage
         public string CurrentPassword { get; set; }
 
         public string NewPassword { get; set; }
+    }
+
+    public class ChangePasswordRequestValidator : AbstractValidator<ChangePasswordRequest>
+    {
+        public ChangePasswordRequestValidator()
+        {
+            RuleFor(x => x.CurrentPassword).NotEmpty().WithName("Current Password");
+            RuleFor(x => x.NewPassword).NotEmpty().MinimumLength(8).MaximumLength(50).WithName("New Password");
+        }
     }
 }
