+++ src/Halcyon.Api/Models/Manage/UpdateProfileRequest.cs
@@ -1,3 +1,7 @@
+using FluentValidation;
+using Halcyon.Api.Services.Date;
+using Halcyon.Api.Validators;
+
 namespace Halcyon.Api.Models.Manage
 {
     public class UpdateProfileRequest : UpdateRequest
@@ -9,5 +13,16 @@ namespace Halcyon.Api.Models.Manage
         public string LastName { get; set; }
 
         public DateOnly? DateOfBirth { get; set; }
+    }
+
+    public class UpdateProfileRequestValidator : AbstractValidator<UpdateProfileRequest>
+    {
+        public UpdateProfileRequestValidator(IDateTimeProvider dateTimeProvider)
+        {
+            RuleFor(x => x.EmailAddress).NotEmpty().EmailAddress().MaximumLength(255).WithName("Email Address");
+            RuleFor(x => x.FirstName).NotEmpty().MaximumLength(50).WithName("First Name");
+            RuleFor(x => x.LastName).NotEmpty().MaximumLength(50).WithName("Last Name");
+            RuleFor(x => x.DateOfBirth).NotEmpty().InThePast(dateTimeProvider).WithName("Date Of Birth");
+        }
     }
 }
