+++ src/Halcyon.Api/Models/User/CreateUserRequest.cs
@@ -1,3 +1,7 @@
+using FluentValidation;
+using Halcyon.Api.Services.Date;
+using Halcyon.Api.Validators;
+
 namespace Halcyon.Api.Models.User
 {
     public class CreateUserRequest
@@ -13,5 +17,17 @@ namespace Halcyon.Api.Models.User
         public DateOnly? DateOfBirth { get; set; }
 
         public List<string> Roles { get; set; }
+    }
+
+    public class CreateUserRequestValidator : AbstractValidator<CreateUserRequest>
+    {
+        public CreateUserRequestValidator(IDateTimeProvider dateTimeProvider)
+        {
+            RuleFor(x => x.EmailAddress).NotEmpty().EmailAddress().MaximumLength(255).WithName("Email Address");
+            RuleFor(x => x.Password).NotEmpty().MinimumLength(8).MaximumLength(50).WithName("Password");
+            RuleFor(x => x.FirstName).NotEmpty().MaximumLength(50).WithName("First Name");
+            RuleFor(x => x.LastName).NotEmpty().MaximumLength(50).WithName("Last Name");
+            RuleFor(x => x.DateOfBirth).NotEmpty().InThePast(dateTimeProvider).WithName("Date Of Birth");
+        }
     }
 }
