+++ src/Halcyon.Api/Models/User/SearchUsersRequest.cs
@@ -1,7 +1,16 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using FluentValidation;
+using System.ComponentModel.DataAnnotations;
 
 namespace Halcyon.Api.Models.User
 {
+    public enum UserSort
+    {
+        EMAIL_ADDRESS_ASC,
+        EMAIL_ADDRESS_DESC,
+        NAME_ASC,
+        NAME_DESC
+    }
+
     public class SearchUsersRequest
     {
         public string Search { get; set; }
@@ -13,5 +22,15 @@ namespace Halcyon.Api.Models.User
 
         [Range(1, 50)]
         public int Size { get; set; } = 50;
+    }
+
+    public class SearchUsersRequestValidator : AbstractValidator<SearchUsersRequest>
+    {
+        public SearchUsersRequestValidator()
+        {
+            RuleFor(x => x.Sort).IsInEnum();
+            RuleFor(x => x.Page).GreaterThanOrEqualTo(1).LessThanOrEqualTo(int.MaxValue);
+            RuleFor(x => x.Size).GreaterThanOrEqualTo(1).LessThanOrEqualTo(50);
+        }
     }
 }
