--- src/Halcyon.Api/Models/User/SearchUsersRequestValidator.cs
@@ -1,14 +1,0 @@
-﻿using FluentValidation;
-
-namespace Halcyon.Api.Models.User
-{
-    public class SearchUsersRequestValidator : AbstractValidator<SearchUsersRequest>
-    {
-        public SearchUsersRequestValidator()
-        {
-            RuleFor(x => x.Sort).IsInEnum();
-            RuleFor(x => x.Page).GreaterThanOrEqualTo(1).LessThanOrEqualTo(int.MaxValue);
-            RuleFor(x => x.Size).GreaterThanOrEqualTo(1).LessThanOrEqualTo(50);
-        }
-    }
-}
