--- src/Halcyon.Api/Models/User/UserSort.cs
@@ -1,10 +1,0 @@
-﻿namespace Halcyon.Api.Models.User
-{
-    public enum UserSort
-    {
-        EMAIL_ADDRESS_ASC,
-        EMAIL_ADDRESS_DESC,
-        NAME_ASC,
-        NAME_DESC
-    }
-}
